# boot/src/main/resources/contracts/swagger.yaml
openapi: 3.0.3
info:
  title: Centimo API
  description: |
    API REST de Centimo — dashboard de finanzas personales.
    Gestiona plataformas, cuentas, snapshots mensuales, gastos, ingresos,
    inversiones, crowdlending, fondos, distribución de sueldo, compromisos y alertas.
  version: 1.0.0
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /nomina/{year}/{month}:
    get:
      tags: [ Nomina ]
      operationId: getNominaAndDate
      summary: Obtener la nomina única en un mes concreto
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
          example: 2026
        - name: month
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
          example: 7
      responses:
        '200':
          description: Nomina encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NominaResponse'
        '404':
          description: No existe nomina en el mes indicado
  /nomina:
    post:
      tags: [ Nomina ]
      operationId: createNomina
      summary: Crear nomina
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NominaCreate'
      responses:
        '201':
          description: Creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NominaResponse'
        '400':
          description: Error generico 400
        '404':
          description: Error generico 404


  # ════════════════════════════════════════════
  #  SEGUIMIENTO MENSUAL
  # ════════════════════════════════════════════

  # ── Snapshots ──
  /snapshots/accounts/{accountId}/{year}/{month}:
    get:
      tags: [ Snapshots ]
      operationId: getSnapshotByAccountAndDate
      summary: Obtener la instantánea única de una cuenta en un mes concreto
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
          example: "acc_bbva_1"
        - name: year
          in: path
          required: true
          schema:
            type: integer
          example: 2026
        - name: month
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
          example: 7
      responses:
        '200':
          description: Instantánea encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotResponse'
        '404':
          description: No existe instantánea para esa cuenta en el mes indicado
  /snapshots:
    post:
      tags: [Snapshots]
      operationId: createSnapshot
      summary: Crear snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MonthlySnapshotCreate'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotResponse'
        '400':
          description: Error generico 400
        '404':
          description: Error generico 404
...
components:
...
  schemas:
...
    NominaResponse:
      type: object
      required: [ id, year, month, value ]
      properties:
        id:
          type: string
        year:
          type: integer
        month:
          type: integer
          minimum: 1
          maximum: 12
        value:
          type: number
          format: decimal
        notes:
          type: string
          nullable: true
    NominaCreate:
      type: object
      required: [ year, month, value ]
      properties:
        year:
          type: integer
        month:
          type: integer
          minimum: 1
          maximum: 12
        value:
          type: number
          format: decimal
        notes:
          type: string
          nullable: true

    # ═══════════ MonthlySnapshot ═══════════
    SnapshotResponse:
      type: object
      required: [id, accountId, year, month, balance, income, expenses]
      properties:
        id:
          type: string
        accountId:
          type: string
        year:
          type: integer
        month:
          type: integer
          minimum: 1
          maximum: 12
        balance:
          type: number
          format: decimal
        income:
          type: number
          format: decimal
        expenses:
          type: number
          format: decimal
        contribution:
          type: number
          format: decimal
          nullable: true
        notes:
          type: string
          nullable: true
        checklistItems:
          type: array
          items:
            $ref: '#/components/schemas/ChecklistItem'
          nullable: true
    MonthlySnapshotCreate:
      type: object
      required: [accountId, year, month, balance, income, expenses]
      properties:
        accountId:
          type: string
        year:
          type: integer
        month:
          type: integer
        balance:
          type: number
          format: float
        income:
          type: number
          format: float
        expenses:
          type: number
          format: float
        contribution:
          type: number
          format: float
          nullable: true
        notes:
          type: string
          nullable: true
        checklistItems:
          type: array
          items:
            $ref: '#/components/schemas/ChecklistItem'
          nullable: true
...
    ChecklistItem:
      type: object
      required: [id, text, checked]
      properties:
        id:
          type: string
        text:
          type: string
        checked:
          type: boolean

    # ═══════════ Expense ═══════════
...